pyproject.toml
[project]
name = "polls-site"
version = "0.1.0"
description = "A Django site running the polls app."
readme = "README.md"
requires-python = ">=3.12,<4.0"

[tool.poetry.dependencies]
django = ">=5.2.5,<6.0.0"
gunicorn = ">=23.0.0,<24.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
polls-app = { git = "https://github.com/akitorahayashi/polls-app.git" }
dj-database-url = ">=2.2.0,<3.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
filelock = ">=3.12.0,<4.0.0"
ipython = "^9.4.0"
pytest = ">=8.0.0,<9.0.0"
pytest-django = ">=4.0.0,<5.0.0"
pytest-xdist = ">=3.0.0,<4.0.0"
ruff = "^0.12.9"
testcontainers = {extras = ["postgres"], version = "^4.6.0"}
httpx = "^0.27.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
markers = [
    "e2e: marks tests as end-to-end tests",
]

[tool.black]
target-version = ['py312']
exclude = '''
/(\
    \.git
    | \.mypy_cache
    | \.venv
    | build
    | dist
)/
'''

[tool.ruff]
target-version = "py312"

exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "build",
    "dist",
	"migrations",
	"db.sqlite3",
	"*.pyc"
]

[tool.ruff.lint]
select = ["E", "F", "I"]
# E501 (line too long) is ignored
ignore = ["E501"]